blogien skeeman määrittely siirretty models/blog
