+++ osa4/blogilista/index.js
@@ -4,14 +4,7 @@ const mongoose = require('mongoose')
 
 const app = express()
 
-const blogSchema = mongoose.Schema({
-  title: String,
-  author: String,
-  url: String,
-  likes: Number,
-})
-
-const Blog = mongoose.model('Blog', blogSchema)
+const Blog = require('./models/blog')
 
 // const password = process.argv[2]
 const mongoUrl = process.env.MONGODB_URI
