+++ osa4/blogilista/models/blog.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose')
+
+const blogSchema = mongoose.Schema({
+  title: String,
+  author: String,
+  url: String,
+  likes: Number,
+})
+
+module.exports = mongoose.model('Blog', blogSchema)
