+++ osa4/blogilista/requests/create_file.rest
@@ -2,8 +2,8 @@ POST http://localhost:3003/api/blogs
 Content-Type: application/json
 
 {
-  "title": "My second  blog",
+  "title": "My third  blog",
   "author": "SV",
-  "url": "http://google.com",
-  "likes": 10
+  "url": "http://facebook.com",
+  "likes": 100
 }
